backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv

# --- LLM / LangChain stack (pinned to keep imports stable) ---
groq==0.12.0
langchain==0.2.11
langchain-core==0.2.23
langchain-community==0.2.11
langchain-groq==0.1.6

# --- Search tool (Google CSE) ---
google-api-python-client

# --- Misc runtime deps ---
aiohttp
python-multipart